Improve all of the gateway code

+++ src/main/kotlin/com/viral32111/discordrelay/config/HTTP.kt
@@ -8,5 +8,5 @@ import kotlinx.serialization.Serializable
 data class HTTP(
 	@Required @SerialName( "user-agent-prefix" ) val userAgentPrefix: String = "Minecraft Server (https://example.com; contact@example.com)",
 	@Required @SerialName( "from-address" ) val fromAddress: String = "contact@example.com",
-	@SerialName( "timeout-seconds" ) val timeoutSeconds: Int = 5
+	@SerialName( "timeout-seconds" ) val timeoutSeconds: Long = 5
 )
